Remove Facebook and OAuth login
Those methods probably don't work anymore.
They only make sense when the competition is held in person and only registered
participants have access to the site. In the online setting we just send them a
username and password. This can also be done in-person.

Optimally we would use something like OAuth and allow only people with a
whitelisted email address to register themselves. But that implementation would
take some effort that we can't spend for this year.