Skip instance if it is unavailable
When trying to run a command on a docker instance, first check if it's
available. If it doesn't respond to `docker info`, mark it as unavailable and
try another one.
This will cause the number of available instances to decrease, without marking
them as available again. Therefore the admin has to keep an eye on the
available instances and restart and manually initiate a re-check of
availability from the admin interface.